+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Phantom Bloom
+- Desolate Talus
+- K'areshi Lotus
+- K'areshi Resonating Stone
+- Manaforged Instrument
+- (Fractured) Spark of Starlight
+
 ## [1.3.3] 2025-08-06
 ### Changed
 - Bump TOC to 11.2.0